+++ tests/testthat/test_nfFileCreation.R
@@ -2,6 +2,7 @@ library(CRSSIO)
 context('check that Natural Flow files are created correctly.')
 
 dir.create('tmp')
+on.exit(unlink('tmp',recursive = T))
 p1 <- '..'
 rr <- sample(1:29, 4) # get 4 random nodes
 message(cat('4 random nodes are:',rr))
@@ -12,6 +13,17 @@ test_that('can create files',{
   expect_message(createCRSSDNFInputFiles('CoRiverNF', oFolder = 'tmp', 
                                          startDate = '2017-1-31', simYrs = 5, 
                                          recordToUse = c('1950-01','1954-12')))
+})
+
+allFiles <- c(CRSSNFInputNames(), "MWD ICS.SacWYType", 
+              "MeadFloodControlData.hydrologyIncrement", "HydrologyParameters.TraceNumber",
+              "HydrologyParameters.SupplyScenario")
+
+test_that("all files exist", {
+  expect_true(all(allFiles %in% list.files("tmp/trace1")))
+  expect_true(all(allFiles %in% list.files("tmp/trace3")))
+  expect_true(all(list.files("tmp/trace1") %in% allFiles))
+  expect_true(all(list.files("tmp/trace3") %in% allFiles))
 })
 
 test_that('files created from "CoRiverNF" are the same as from Excel', {
@@ -34,4 +46,3 @@ test_that('ism files match each other as expected', {
                as.matrix(read.csv(file.path(p1,'trace5/', CRSSNFInputNames()[rr[2]]),skip=1))[1:12])
 })
 
-unlink('tmp',recursive = T) # delete the files that were created in the tests