+++ NEWS.md
@@ -1,4 +1,22 @@
+# CRSSIO 0.5.0
+
+*Released August 30, 2017* 
+
+## New Features and Functions
+
+* `createISMMatrix()` now works with annual data as well as monthly data. (#24)
+* `createCRSSDNFInputFiles()` now creates Sacramento year type index using ISM and saves it (#25)
+* new function: `elevation2Volume()` that will convert reservoir elevation values to resevoir storage values
+
+## Minor Updates and Bug Fixes
+
+* `createSysCondTable()` now checks to see if `yrs` exist in the data and reacts properly. (#26)
+* package now depends on xts >= 0.10.0 (#17)
+* Improved unit tests
+
 # CRSSIO 0.4.1
+
+*Released May 10, 2017*
 
 ## Bug Fixes
 
@@ -6,7 +24,7 @@
 
 # CRSSIO 0.4.0
 
-Released March 31, 2017
+*Released March 31, 2017*
 
 ## Major new features